apps/api/requirements.txt
# ===========================================
# NEZUKO ADMIN API - PYTHON DEPENDENCIES
# Last Updated: January 24, 2026
# ===========================================

# Core Framework
fastapi>=0.124.4
uvicorn[standard]>=0.40.0
python-multipart>=0.0.20

# Authentication
python-jose[cryptography]>=3.5.0
passlib[argon2]>=1.7.4

# Validation
pydantic>=2.12.5
pydantic-settings>=2.7.0
email-validator>=2.2.0

# Database
sqlalchemy[asyncio]>=2.0.46
asyncpg>=0.31.0
alembic>=1.18.1

# Cache
redis>=5.2.1

# HTTP & WebSocket
httpx>=0.28.1
websockets>=14.1

# Logging
structlog>=25.1.0

# Monitoring
prometheus-client>=0.21.0
sentry-sdk[fastapi]>=2.19.0

# Rate Limiting
slowapi>=0.1.9
